third_party/WebKit/LayoutTests/platform/chromium-mac/http/tests/xmlhttprequest/XMLHttpRequestException-expected.txt

Tests the properties of XMLHttpRequest exceptions.

PASS e.toString() is "Error: NetworkError: DOM Exception 19"
PASS Object.prototype.toString.call(e) is "[object DOMException]"
FAIL Object.prototype.toString.call(e.__proto__) is not "[object DOMException]". Was "[object Object]"
FAIL e.constructor.toString() is not "[object DOMExceptionConstructor]". Was "function DOMException() { [native code] }"
PASS e.constructor is window.DOMException
PASS e.NETWORK_ERR is e.constructor.NETWORK_ERR
PASS e.NETWORK_ERR is 19
